backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.21
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
